✨ Config for LR search experiment for 1.3B
We found that our estimate for optimal LR may be off by 1 order. This set of experiments (wandb id: al78e4be) explored a higher LR of 1e-3. This LR seems to be better than the 1e-4 tried earlier. I also tried 1e-2 but that diverges quickly.

Also discovered that for FSDP + autoresume to work, the checkpoint either needs to be sharded or load_monolith_rank0_only needs to be set. Otherwise the ranks get an OOM error by trying to load the full model on each GPU. Also discovered that the  flag for use_original_params=False makes MFU more steady and overall about 5% higher than without the flag.